// src/ea-langgraph-sdk/types.d.ts
export interface LangGraphAgentState<TContext extends Record<string, unknown> = Record<string, unknown>> {
    agentRuntime: AgentRuntimeServiceApi;
    context: TContext;
}